created basic myaccount page
