+++ client/Routes.js
@@ -4,7 +4,7 @@ import { Route, Switch, Redirect } from 'react-router-dom'
 import { Login, Signup } from './components/AuthForm'
 import AllProducts from './components/AllProducts'
 import Cart from './components/Cart'
-import UserOption from './components/UserOption'
+import MyAccount from './components/MyAccount'
 import AdminBoard from './components/AdminBoard'
 import { me } from './store'
 import Checkout from './components/Checkout'
@@ -24,8 +24,8 @@ const Routes = () => {
     <div>
       {!!isLoggedIn ? (
         <Switch>
-          <Route exact path="/useroption">
-            <UserOption />
+          <Route exact path="/myaccount">
+            <MyAccount />
           </Route>
           <Route exact path="/products">
             <AllProducts />
@@ -39,7 +39,7 @@ const Routes = () => {
           <Route path="/checkout">
             <Checkout />
           </Route>
-          <Route path="/adminOptions">
+          <Route exact path="/adminOptions">
             <AdminOptions />
           </Route>
         </Switch>
